+++ CHANGES.md
@@ -3,6 +3,7 @@
 
 - Fontify Ansible documentation
 - Add buttons to follow references to other modules
+- Add support for imenu (currently options only)
 - Add support for bookmarks on Ansible documentation
 - Allow input even if no Ansible modules were found